requirements.txt
django>=3.2,<3.3
djangorestframework
pillow